// src/common/enums.ts
// Local storage Keys
export enum LOCAL_KEY {
  ACCESS_TOKEN = "access-token",
  USER = "user",
  ORDER_ID = "ORDER_ID",
}

export enum ServiceTypeEnum {
  Home = "home",
  Flight = "flight",
  Hotel = "hotel",
  Bus = "bus",
  Shop = "shop",
  Gigs = "gigs",
}

export enum FlightSearchTypeEnum {
  "Round Trip",
  "One Way",
}